refactor(search): slice 1 — extract result/options/restarts from solve.rs
First slice of the §6 type split. Pure file split: no API, behavior,
or import-path changes for external consumers. `search/solve.rs`
re-exports the moved types under their original `solve::*` paths so
internal `use crate::search::solve::SolveStatus` (etc.) keeps
resolving until the migration completes.

Files:

- `search/result.rs` (new) — `SolveStatus`, `SolveResult` (with the
  `solved` / `unsolved` / `unsolvable` constructors).
- `search/options.rs` (new) — `Strategy`, `InnerStrategy`,
  `SolveOptions` (with `upgraded_for_entangling`),
  `EntropyOptions`, `EntanglingOptions` (with `clipped_future_layers`).
- `search/restarts.rs` (new) — `extract`, `pick_best`,
  `run_with_components`, `run_strategy_v2`. All `pub(crate)` —
  internal infrastructure for `MoveSolver`.
- `search/solve.rs` — keeps `MoveSolver` + `CandidateAttempt` +
  `MultiSolveResult` + the `EntanglingCache` / `NoHomeCache`
  arch-derived caches that `MoveSolver` owns. Drops 595 lines
  (2025 → 1430).

`search/mod.rs` declares the three new submodules and the existing
`solve.rs`. The doc comment now lists the slice-1 layout and previews
the forthcoming `move_search.rs` / `target_solver.rs` siblings (§6).

Verified:

- 284 search-crate unit tests pass.
- 60-case `temp_regression` integration suite passes bit-for-bit on
  status / cost / nodes_expanded / deadlocks / goal_config /
  move_layers.
- `cargo clippy -p bloqade-lanes-search --all-targets -- -D warnings`
  is clean.
