+++ crates/bloqade-lanes-search/src/search/solve.rs
@@ -3,6 +3,11 @@
 //!
 //! [`MoveSolver`] is constructed once per architecture (parsing JSON and
 //! building indexes) and can then solve multiple placement problems.
+//!
+//! Slice-1 of the §6 type split has factored the result types, options
+//! bundles, and restart orchestration into [`super::result`],
+//! [`super::options`], and [`super::restarts`]; the next slices will
+//! extract `MoveSearch` / `TargetSolver` and the `CzPlacement` peers.
 
 use std::collections::{HashMap, HashSet};
 use std::sync::{Arc, OnceLock};
@@ -11,626 +16,26 @@ use bloqade_lanes_bytecode_core::arch::addr::LocationAddr;
 use bloqade_lanes_bytecode_core::arch::types::ArchSpec;
 use rayon::prelude::*;
 
-use crate::cost::UniformCost;
-use crate::drivers::astar::SearchResult;
-use crate::drivers::entropy::EntropyTrace;
-use crate::drivers::frontier::{self, BfsFrontier, DfsFrontier, IdsFrontier, PriorityFrontier};
 use crate::generators::HeuristicGenerator;
 use crate::generators::heuristic::DeadlockPolicy;
 use crate::goals::AllAtTarget;
-use crate::observer::NoOpObserver;
 use crate::ops::entangling::{self, WordPairDistances};
-use crate::ops::entangling::{LOOKAHEAD_BETA, MOVE_PENALTY, OCCUPANCY_PENALTY_DEFAULT};
+use crate::ops::entangling::{LOOKAHEAD_BETA, MOVE_PENALTY};
 use crate::placement::nohome::{self, NoHomeOptions};
 use crate::placement::target_generator::{TargetContext, TargetGenerator, validate_candidate};
 use crate::primitives::config::{Config, ConfigError};
-use crate::primitives::context::{SearchContext, SearchState};
+use crate::primitives::context::SearchContext;
 use crate::primitives::distance::{DistanceTable, HopDistanceHeuristic};
-use crate::primitives::graph::MoveSet;
 use crate::primitives::lane_index::LaneIndex;
-use crate::scorers::DistanceScorer;
-use crate::traits::{Goal, Heuristic, MoveGenerator};
+use crate::search::restarts::{pick_best, run_with_components};
 
-/// Inner strategy for the cascade's Phase 1 (fast feasibility search).
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum InnerStrategy {
-    /// Iterative Diving Search.
-    Ids,
-    /// Heuristic depth-first search.
-    Dfs,
-    /// Entropy-guided search.
-    Entropy,
-}
-
-/// Search strategy for the solver.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum Strategy {
-    /// A* / Weighted A* search: `f = g + weight * h`, goal on pop.
-    /// weight=1.0 is standard A* (optimal); weight>1.0 is bounded suboptimal.
-    AStar,
-    /// Heuristic depth-first search: fast, bounded memory, not optimal.
-    HeuristicDfs,
-    /// Breadth-first search: finds shallowest solution, no heuristic.
-    Bfs,
-    /// Greedy best-first: fast, uses heuristic only (no path cost).
-    GreedyBestFirst,
-    /// Iterative Diving Search: depth-first with heuristic jump-back.
-    Ids,
-    /// Cascade: fast inner strategy first, then weighted A* bounded by inner cost.
-    /// Restarts apply to the inner phase only; A* runs once with the tightest bound.
-    Cascade { inner: InnerStrategy },
-    /// Entropy-guided search: single-path DFS with entropy-based backtracking.
-    Entropy,
-}
-
-/// Outcome status of a solve attempt.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum SolveStatus {
-    /// A solution was found.
-    Solved,
-    /// The search space was fully explored — no solution exists.
-    Unsolvable,
-    /// The expansion budget was exhausted before finding a solution or
-    /// proving unsolvability.
-    BudgetExceeded,
-}
-
-impl SolveStatus {
-    /// Stable string label for status reporting (PyO3 wrappers, logs).
-    pub fn as_label(&self) -> &'static str {
-        match self {
-            Self::Solved => "solved",
-            Self::Unsolvable => "unsolvable",
-            Self::BudgetExceeded => "budget_exceeded",
-        }
-    }
-}
-
-/// Result of a solve attempt.
-///
-/// Always returned (never `None`). Check [`status`](SolveResult::status) to
-/// determine whether a solution was found.
-#[derive(Debug)]
-pub struct SolveResult {
-    /// Whether the solve succeeded, was unsolvable, or ran out of budget.
-    pub status: SolveStatus,
-    /// Sequence of parallel move steps from initial to goal configuration.
-    /// Empty when `status` is not `Solved`.
-    pub move_layers: Vec<MoveSet>,
-    /// Final qubit positions (the goal configuration).
-    /// Equals the initial configuration when `status` is not `Solved`.
-    pub goal_config: Config,
-    /// Number of nodes expanded during search.
-    pub nodes_expanded: u32,
-    /// Total path cost. 0.0 when `status` is not `Solved`.
-    pub cost: f64,
-    /// Number of deadlocks encountered during search.
-    pub deadlocks: u32,
-    /// Optional entropy-search trace payload for visualization/debugging.
-    pub entropy_trace: Option<EntropyTrace>,
-}
-
-impl SolveResult {
-    /// Construct a [`SolveStatus::Solved`] result with the given path and counters.
-    pub fn solved(
-        goal_config: Config,
-        move_layers: Vec<MoveSet>,
-        cost: f64,
-        nodes_expanded: u32,
-        deadlocks: u32,
-    ) -> Self {
-        Self {
-            status: SolveStatus::Solved,
-            move_layers,
-            goal_config,
-            nodes_expanded,
-            cost,
-            deadlocks,
-            entropy_trace: None,
-        }
-    }
-
-    /// Construct a non-Solved result anchored at `root_config`.
-    /// Use [`SolveStatus::Unsolvable`] or [`SolveStatus::BudgetExceeded`].
-    /// `move_layers` is empty and `cost` is `0.0` by definition.
-    pub fn unsolved(
-        status: SolveStatus,
-        root_config: Config,
-        nodes_expanded: u32,
-        deadlocks: u32,
-    ) -> Self {
-        debug_assert!(
-            !matches!(status, SolveStatus::Solved),
-            "SolveResult::unsolved cannot carry SolveStatus::Solved; use SolveResult::solved"
-        );
-        Self {
-            status,
-            move_layers: Vec::new(),
-            goal_config: root_config,
-            nodes_expanded,
-            cost: 0.0,
-            deadlocks,
-            entropy_trace: None,
-        }
-    }
-
-    /// Convenience for [`SolveStatus::Unsolvable`] with zero counters
-    /// (no expansions happened, no deadlocks encountered).
-    pub fn unsolvable(root_config: Config) -> Self {
-        Self::unsolved(SolveStatus::Unsolvable, root_config, 0, 0)
-    }
-}
-
-/// Core search-tuning parameters shared by every [`MoveSolver`] entry point.
-///
-/// Strategy-specific knobs live in [`EntropyOptions`] (entropy-search
-/// parameters) and [`EntanglingOptions`] (loose-goal Hungarian parameters).
-/// Problem-specific data (`initial`, `target`, `blocked`, `max_expansions`)
-/// remain as direct arguments.
-#[derive(Debug, Clone)]
-pub struct SolveOptions {
-    /// Search strategy to use.
-    pub strategy: Strategy,
-    /// Heuristic weight for A* (1.0 = standard, >1.0 = bounded suboptimal).
-    pub weight: f64,
-    /// Number of parallel restarts with perturbed scoring (1 = no restarts).
-    pub restarts: u32,
-    /// How to handle deadlocks (no improving moves).
-    pub deadlock_policy: DeadlockPolicy,
-    /// Enable 2-step lookahead scoring inside [`HeuristicGenerator`].
-    /// Affects every strategy that goes through that generator.
-    pub lookahead: bool,
-    /// Per-qubit move-candidate pruning passed to [`HeuristicGenerator`].
-    /// `None` keeps all scored triples (default for the basic [`solve`]
-    /// path); `Some(n)` keeps the top `n` bus options per qubit by score.
-    /// [`solve_entangling`] defaults this to `Some(3)` when not set.
-    pub top_c: Option<usize>,
-}
-
-impl Default for SolveOptions {
-    fn default() -> Self {
-        Self {
-            strategy: Strategy::AStar,
-            weight: 1.0,
-            restarts: 1,
-            deadlock_policy: DeadlockPolicy::Skip,
-            lookahead: false,
-            top_c: None,
-        }
-    }
-}
-
-/// Entropy-strategy-specific parameters.
-///
-/// Only consumed when [`SolveOptions::strategy`] is [`Strategy::Entropy`]
-/// (or a [`Strategy::Cascade`] variant whose inner is entropy). Pass via
-/// [`MoveSolver::solve`]'s optional `entropy_opts` argument; otherwise
-/// defaults are used.
-#[derive(Debug, Clone)]
-pub struct EntropyOptions {
-    /// Max movesets generated per bus group.
-    pub max_movesets_per_group: usize,
-    /// Number of goal candidates to collect before stopping entropy search.
-    pub max_goal_candidates: usize,
-    /// Time-distance blend weight (0.0 = hop-count only, 1.0 = time only).
-    pub w_t: f64,
-    /// Collect entropy-step trace payload for visualization/debugging.
-    pub collect_entropy_trace: bool,
-}
-
-impl Default for EntropyOptions {
-    fn default() -> Self {
-        Self {
-            max_movesets_per_group: 3,
-            max_goal_candidates: 3,
-            w_t: 0.05,
-            collect_entropy_trace: false,
-        }
-    }
-}
-
-/// Loose-goal entangling-search parameters consumed by
-/// [`MoveSolver::solve_entangling`].
-///
-/// Ignored by [`MoveSolver::solve`] and [`MoveSolver::solve_nohome`].
-#[derive(Debug, Clone)]
-pub struct EntanglingOptions {
-    /// Congestion penalty weight for the entangling Hungarian assignment.
-    ///
-    /// `0.0` (default): standard min-sum-distance assignment.
-    /// `> 0.0`: iteratively re-runs Hungarian, adding a penalty
-    /// `congestion_weight × (load - ideal_load)` to slots on overloaded
-    /// entangling word pairs. Spreads CZ pairs across word pairs to reduce
-    /// routing serialization at high occupancy.
-    pub congestion_weight: f64,
-    /// Spectator-occupancy penalty (in lane-hop units) added to each
-    /// entangling Hungarian cost cell *per spectator-occupied slot half*.
-    /// A spectator is an atom that is *not* in any CZ pair of the current
-    /// layer.
-    ///
-    /// `0.0`: occupancy-blind (legacy behaviour). `> 0.0`: bias the
-    /// assignment away from slots that would force the search to evict a
-    /// non-participating atom. Default `1.0` was selected by sweep on
-    /// the 80q / depth 3 / max_pairs 10 regime; deeper sparse-pair
-    /// circuits prefer larger values (~2–3), full-layer circuits are
-    /// unaffected since they have no spectators. Atoms that are
-    /// themselves part of another CZ pair this layer are *not* penalised
-    /// — they will be reassigned by the Hungarian and move out of the
-    /// way naturally.
-    ///
-    /// Internally the per-cell contribution is
-    /// `(spectator_half_count * occupancy_penalty).round() as u32`, so
-    /// fractional values (e.g. `0.5`, `1.5`) provide meaningful sub-hop
-    /// granularity for sweeps. Must be finite and non-negative.
-    pub occupancy_penalty: f64,
-    /// Cap on the number of future CZ layers fed to the Hungarian
-    /// forward/backward sweep. `None` is unbounded; `Some(0)` disables
-    /// lookahead entirely (single-layer Hungarian); `Some(n)` for n > 0
-    /// keeps the first `n` future layers. Default `Some(4)` keeps solve
-    /// time bounded regardless of circuit depth.
-    pub hungarian_horizon: Option<usize>,
-}
-
-impl Default for EntanglingOptions {
-    fn default() -> Self {
-        Self {
-            congestion_weight: 0.0,
-            occupancy_penalty: OCCUPANCY_PENALTY_DEFAULT,
-            hungarian_horizon: Some(4),
-        }
-    }
-}
-
-impl EntanglingOptions {
-    /// Apply this options bundle's Hungarian future-layer horizon to a
-    /// list of upcoming CZ layers.
-    ///
-    /// `hungarian_horizon == None` is unbounded; `Some(0)` disables
-    /// multi-layer lookahead entirely; `Some(n > 0)` keeps the first
-    /// `n` layers.
-    pub fn clipped_future_layers<'a>(
-        &self,
-        future_cz_layers: &'a [Vec<(u32, u32)>],
-    ) -> &'a [Vec<(u32, u32)>] {
-        match self.hungarian_horizon {
-            Some(0) => &[],
-            Some(n) => &future_cz_layers[..future_cz_layers.len().min(n)],
-            None => future_cz_layers,
-        }
-    }
-}
-
-impl SolveOptions {
-    /// Upgrade [`DeadlockPolicy::Skip`] to [`DeadlockPolicy::MoveBlockers`]
-    /// for entangling-style routing — the loose-goal solver needs at
-    /// least `MoveBlockers` to handle qubits competing for entangling
-    /// positions. `MoveBlockers` and `AllMoves` pass through unchanged.
-    pub fn upgraded_for_entangling(&self) -> SolveOptions {
-        if matches!(self.deadlock_policy, DeadlockPolicy::Skip) {
-            SolveOptions {
-                deadlock_policy: DeadlockPolicy::MoveBlockers,
-                ..self.clone()
-            }
-        } else {
-            self.clone()
-        }
-    }
-}
-
-// ── Shared helpers ────────────────────────────────────────────────
-
-/// Extract a [`SolveResult`] from a [`SearchResult`].
-fn extract(result: SearchResult, deadlocks: u32, max_exp: Option<u32>) -> SolveResult {
-    match result.goal {
-        Some(goal_id) => {
-            let move_layers = result.solution_path().unwrap_or_default();
-            let goal_config = result.graph.config(goal_id).clone();
-            let cost = result.graph.g_score(goal_id);
-            SolveResult::solved(
-                goal_config,
-                move_layers,
-                cost,
-                result.nodes_expanded,
-                deadlocks,
-            )
-        }
-        None => {
-            let root_config = result.graph.config(result.graph.root()).clone();
-            let status = if max_exp.is_some_and(|max| result.nodes_expanded >= max) {
-                SolveStatus::BudgetExceeded
-            } else {
-                SolveStatus::Unsolvable
-            };
-            SolveResult::unsolved(status, root_config, result.nodes_expanded, deadlocks)
-        }
-    }
-}
-
-/// Pick the best result from multiple restarts (prefer solved, then lowest cost).
-fn pick_best(results: Vec<SolveResult>) -> SolveResult {
-    results
-        .into_iter()
-        .min_by(|a, b| {
-            let a_solved = a.status == SolveStatus::Solved;
-            let b_solved = b.status == SolveStatus::Solved;
-            b_solved.cmp(&a_solved).then(a.cost.total_cmp(&b.cost))
-        })
-        .expect("non-empty results")
-}
-
-/// Shared strategy dispatch + restart logic.
-///
-/// Both [`MoveSolver::solve`] and [`MoveSolver::solve_entangling`] delegate
-/// here after constructing their specific goal, heuristic, and generator.
-#[allow(clippy::too_many_arguments)]
-fn run_with_components<Go, Gen, Hmax, Hsum, MkGen>(
-    root: Config,
-    goal: &Go,
-    make_generator: MkGen,
-    h_max: Hmax,
-    h_sum: Hsum,
-    ctx: &SearchContext,
-    max_expansions: Option<u32>,
-    opts: &SolveOptions,
-    entropy_opts: Option<&EntropyOptions>,
-) -> SolveResult
-where
-    Go: Goal + Sync,
-    Gen: MoveGenerator,
-    Hmax: Heuristic + Copy + Sync,
-    Hsum: Heuristic + Copy + Sync,
-    MkGen: Fn(u64, DeadlockPolicy) -> Gen + Sync,
-{
-    let strategy = opts.strategy;
-    let weight = opts.weight;
-    let restarts = opts.restarts;
-    let deadlock_policy = opts.deadlock_policy;
-    let entropy_defaults = EntropyOptions::default();
-    let entropy = entropy_opts.unwrap_or(&entropy_defaults);
-    let max_movesets_per_group = entropy.max_movesets_per_group;
-    let max_goal_candidates = entropy.max_goal_candidates;
-    let collect_entropy_trace = entropy.collect_entropy_trace;
-    let w_t = entropy.w_t;
-
-    let scorer = DistanceScorer;
-    let cost_fn = UniformCost;
-
-    // Helper: run a single inner strategy with the given seed and budget.
-    let run_inner = |inner: InnerStrategy, seed: u64, budget: Option<u32>| -> SolveResult {
-        match inner {
-            InnerStrategy::Ids => {
-                let move_gen = make_generator(seed, deadlock_policy);
-                let mut f = IdsFrontier::new(h_sum);
-                let result = frontier::run_search(
-                    root.clone(),
-                    &move_gen,
-                    &scorer,
-                    &cost_fn,
-                    goal,
-                    &mut f,
-                    ctx,
-                    &mut SearchState::default(),
-                    &mut NoOpObserver,
-                    budget,
-                    None,
-                );
-                extract(result, move_gen.deadlock_count(), budget)
-            }
-            InnerStrategy::Dfs => {
-                let move_gen = make_generator(seed, deadlock_policy);
-                let mut f = DfsFrontier::new(h_sum);
-                let result = frontier::run_search(
-                    root.clone(),
-                    &move_gen,
-                    &scorer,
-                    &cost_fn,
-                    goal,
-                    &mut f,
-                    ctx,
-                    &mut SearchState::default(),
-                    &mut NoOpObserver,
-                    budget,
-                    None,
-                );
-                extract(result, move_gen.deadlock_count(), budget)
-            }
-            InnerStrategy::Entropy => {
-                let entropy_params = crate::drivers::entropy::EntropyParams {
-                    max_movesets_per_group,
-                    max_goal_candidates,
-                    lookahead: opts.lookahead,
-                    w_t,
-                    ..crate::drivers::entropy::EntropyParams::default()
-                };
-                let mut entropy_trace = if collect_entropy_trace {
-                    Some(EntropyTrace::for_params(&entropy_params))
-                } else {
-                    None
-                };
-                let result = {
-                    let mut noop = crate::observer::NoOpObserver;
-                    let observer: &mut dyn crate::observer::SearchObserver =
-                        match entropy_trace.as_mut() {
-                            Some(trace) => trace,
-                            None => &mut noop,
-                        };
-                    crate::drivers::entropy::entropy_search(
-                        root.clone(),
-                        goal,
-                        &entropy_params,
-                        ctx,
-                        budget,
-                        None,
-                        seed,
-                        observer,
-                    )
-                };
-                let mut solve = extract(result, 0, budget);
-                solve.entropy_trace = entropy_trace;
-                solve
-            }
-        }
-    };
-
-    // Helper: run inner strategy with parallel restarts, return best.
-    let run_inner_with_restarts = |inner: InnerStrategy| -> SolveResult {
-        if restarts <= 1 {
-            run_inner(inner, 0, max_expansions)
-        } else {
-            let results: Vec<SolveResult> = (0..restarts)
-                .into_par_iter()
-                .map(|i| run_inner(inner, i as u64 + 1, max_expansions))
-                .collect();
-            pick_best(results)
-        }
-    };
-
-    // ── Cascade: inner restarts + single A* refinement ─────────
-    if let Strategy::Cascade { inner } = strategy {
-        let inner_result = run_inner_with_restarts(inner);
-
-        if inner_result.status != SolveStatus::Solved {
-            return inner_result;
-        }
-
-        let max_depth = Some(inner_result.cost.ceil() as u32);
-        let astar_move_gen = make_generator(0, DeadlockPolicy::MoveBlockers);
-        let mut astar_f = PriorityFrontier::astar(h_max, weight);
-        let astar_result = frontier::run_search(
-            root.clone(),
-            &astar_move_gen,
-            &scorer,
-            &cost_fn,
-            goal,
-            &mut astar_f,
-            ctx,
-            &mut SearchState::default(),
-            &mut NoOpObserver,
-            max_expansions,
-            max_depth,
-        );
-        let astar_solve = extract(
-            astar_result,
-            astar_move_gen.deadlock_count(),
-            max_expansions,
-        );
-
-        if astar_solve.status == SolveStatus::Solved {
-            return pick_best(vec![inner_result, astar_solve]);
-        }
-        return inner_result;
-    }
-
-    // ── Non-cascade strategies ─────────────────────────────────
-
-    let run_once = |seed: u64, budget: Option<u32>| -> SolveResult {
-        match strategy {
-            Strategy::Entropy => run_inner(InnerStrategy::Entropy, seed, budget),
-            Strategy::Ids => run_inner(InnerStrategy::Ids, seed, budget),
-            Strategy::HeuristicDfs => run_inner(InnerStrategy::Dfs, seed, budget),
-            _ => {
-                let move_gen = make_generator(seed, DeadlockPolicy::MoveBlockers);
-                let result = run_strategy_v2(
-                    strategy,
-                    root.clone(),
-                    &move_gen,
-                    &scorer,
-                    &cost_fn,
-                    goal,
-                    ctx,
-                    h_max,
-                    budget,
-                    weight,
-                );
-                extract(result, move_gen.deadlock_count(), budget)
-            }
-        }
-    };
-
-    if restarts <= 1 {
-        run_once(0, max_expansions)
-    } else {
-        let results: Vec<SolveResult> = (0..restarts)
-            .into_par_iter()
-            .map(|i| run_once(i as u64 + 1, max_expansions))
-            .collect();
-        pick_best(results)
-    }
-}
-
-/// Dispatch to the appropriate frontier-based search strategy.
-#[allow(clippy::too_many_arguments)]
-fn run_strategy_v2<Go, Gen, Hmax>(
-    strategy: Strategy,
-    root: Config,
-    generator: &Gen,
-    scorer: &DistanceScorer,
-    cost_fn: &UniformCost,
-    goal: &Go,
-    ctx: &SearchContext<'_>,
-    heuristic_fn: Hmax,
-    max_expansions: Option<u32>,
-    weight: f64,
-) -> SearchResult
-where
-    Go: Goal,
-    Gen: MoveGenerator,
-    Hmax: Heuristic + Copy,
-{
-    match strategy {
-        Strategy::AStar => {
-            let mut f = PriorityFrontier::astar(heuristic_fn, weight);
-            frontier::run_search(
-                root,
-                generator,
-                scorer,
-                cost_fn,
-                goal,
-                &mut f,
-                ctx,
-                &mut SearchState::default(),
-                &mut NoOpObserver,
-                max_expansions,
-                None,
-            )
-        }
-        Strategy::Bfs => {
-            let mut f = BfsFrontier::new();
-            frontier::run_search(
-                root,
-                generator,
-                scorer,
-                cost_fn,
-                goal,
-                &mut f,
-                ctx,
-                &mut SearchState::default(),
-                &mut NoOpObserver,
-                max_expansions,
-                None,
-            )
-        }
-        Strategy::GreedyBestFirst => {
-            let mut f = PriorityFrontier::greedy(heuristic_fn);
-            frontier::run_search(
-                root,
-                generator,
-                scorer,
-                cost_fn,
-                goal,
-                &mut f,
-                ctx,
-                &mut SearchState::default(),
-                &mut NoOpObserver,
-                max_expansions,
-                None,
-            )
-        }
-        _ => {
-            unreachable!("IDS/DFS/Cascade/Entropy handled before run_strategy_v2")
-        }
-    }
-}
+// Re-export the moved types under the original `solve::*` path so existing
+// `use crate::search::solve::SolveStatus` imports keep resolving until the
+// consumer migration completes.
+pub use crate::search::options::{
+    EntanglingOptions, EntropyOptions, InnerStrategy, SolveOptions, Strategy,
+};
+pub use crate::search::result::{SolveResult, SolveStatus};
 
 /// Cached architecture-dependent data for [`MoveSolver::solve_entangling`].
 ///
